M8: Loki, Tempo and Pyroscope parity

+++ crates/traces/src/metricsgen/servicegraph/edge_agg.rs
@@ -4,21 +4,21 @@ pub(crate) struct EdgeAgg {
     pub(crate) failed: f64,
     pub(crate) client_seconds_sum: f64,
     pub(crate) client_seconds_count: f64,
-    pub(crate) client_bucket_counts: Vec<u64>,
+    pub(crate) client_bucket_counts: Vec<f64>,
     pub(crate) server_seconds_sum: f64,
     pub(crate) server_seconds_count: f64,
-    pub(crate) server_bucket_counts: Vec<u64>,
+    pub(crate) server_bucket_counts: Vec<f64>,
     pub(crate) messaging_seconds_sum: f64,
     pub(crate) messaging_seconds_count: f64,
-    pub(crate) messaging_bucket_counts: Vec<u64>,
+    pub(crate) messaging_bucket_counts: Vec<f64>,
 }
 
 impl EdgeAgg {
     pub(crate) fn new(bucket_count: usize) -> Self {
         Self {
-            client_bucket_counts: vec![0; bucket_count],
-            server_bucket_counts: vec![0; bucket_count],
-            messaging_bucket_counts: vec![0; bucket_count],
+            client_bucket_counts: vec![0.0; bucket_count],
+            server_bucket_counts: vec![0.0; bucket_count],
+            messaging_bucket_counts: vec![0.0; bucket_count],
             ..Self::default()
         }
     }
